+++ src/scripts/manifest-page.ts
@@ -1,0 +1,3 @@
+import { enhanceDeadlineTimes } from "./deadline-time";
+
+enhanceDeadlineTimes();